Added integration tests. Fixed profile photos

+++ todo/database/migrations/2023_09_27_103948_change_is_completed_column_type_in_todos.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('todos', function (Blueprint $table) {
-            $table->tinyInteger('is_completed')->default(0)->change();
+            $table->integer('is_completed')->default(0)->change();
         });
     }
